# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IAA 投放运营中台 API",
    "description": "IAA 投放运营中台 Java 业务服务契约；示例数据均为虚构。模型暂由 FastAPI 导出，x-implementation 表示 Java 状态；x-reference-implementation 保留 Python 对照服务的实际状态。",
    "version": "1.0.0-rc.1"
  },
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "tags": [
          "用户与权限"
        ],
        "summary": "账号列表",
        "description": "公司管理员看到本租户账号；代理管理员只看到本代理下的账号。",
        "operationId": "listUsers",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "title": "Keyword"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "active",
                    "disabled"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "账号分页",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPage"
                },
                "examples": {
                  "demo": {
                    "summary": "一个运营账号",
                    "value": {
                      "items": [
                        {
                          "id": "usr_01",
                          "username": "demo.operator",
                          "display_name": "演示运营",
                          "roles": [
                            "operator"
                          ],
                          "status": "active",
                          "data_scope": {
                            "products": [
                              "晴空天气",
                              "脑力答题王"
                            ]
                          },
                          "revision": 2
                        }
                      ],
                      "total": 1
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌已失效",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "无权限，或超出账号的数据范围",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-requirements": [
          "IAA-REQ-001",
          "IAA-REQ-003"
        ],
        "x-screens": [
          "user-manage"
        ],
        "x-fields": [
          "username",
          "roles",
          "data_scope",
          "revision"
        ],
        "x-implementation": "implemented",
        "x-reference-implementation": {
          "runtime": "FastAPI",
          "status": "planned"
        },
        "x-implementation-source": "backend-java/src/main/java/com/iaaops/iam/web/UserAdminController.java"
      },
      "post": {
        "tags": [
          "用户与权限"
        ],
        "summary": "创建账号",
        "description": "服务端生成一次性临时口令，只在本响应返回；不能授予超出操作者自身范围的角色或数据。",
        "operationId": "createUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              },
              "examples": {
                "demo": {
                  "summary": "代理管理员",
                  "value": {
                    "username": "demo.agency",
                    "display_name": "演示代理管理员",
                    "roles": [
                      "agency_admin"
                    ],
                    "data_scope": {
                      "agencies": [
                        "星河代理"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "已创建",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCreated"
                },
                "examples": {
                  "demo": {
                    "summary": "创建代理管理员",
                    "value": {
                      "user": {
                        "id": "usr_02",
                        "username": "demo.agency",
                        "display_name": "演示代理管理员",
                        "roles": [
                          "agency_admin"
                        ],
                        "status": "active",
                        "data_scope": {
                          "agencies": [
                            "星河代理"
                          ]
                        },
                        "revision": 1
                      },
                      "one_time_password": "<仅本次返回的临时口令>",
                      "must_change_password": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌已失效",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "无权限，或超出账号的数据范围",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本已变化，请刷新后重试",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-requirements": [
          "IAA-REQ-001",
          "IAA-REQ-003"
        ],
        "x-screens": [
          "user-manage"
        ],
        "x-fields": [
          "username",
          "roles",
          "data_scope",
          "revision"
        ],
        "x-implementation": "implemented",
        "x-reference-implementation": {
          "runtime": "FastAPI",
          "status": "planned"
        },
        "x-implementation-source": "backend-java/src/main/java/com/iaaops/iam/web/UserAdminController.java"
      }
    },
    "/api/v1/users/{user_id}/roles": {
      "put": {
        "tags": [
          "用户与权限"
        ],
        "summary": "修改角色与数据范围",
        "description": "带 revision 防止覆盖他人修改；真实口径与对外口径权限不能同时授予。",
        "operationId": "updateUserRoles",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 64,
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRolesUpdate"
              },
              "examples": {
                "demo": {
                  "summary": "限定产品范围",
                  "value": {
                    "roles": [
                      "operator"
                    ],
                    "data_scope": {
                      "products": [
                        "晴空天气",
                        "脑力答题王"
                      ]
                    },
                    "revision": 2
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "已修改",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSummary"
                },
                "examples": {
                  "demo": {
                    "summary": "改为运营",
                    "value": {
                      "id": "usr_01",
                      "username": "demo.operator",
                      "display_name": "演示运营",
                      "roles": [
                        "operator"
                      ],
                      "status": "active",
                      "data_scope": {
                        "products": [
                          "晴空天气",
                          "脑力答题王"
                        ]
                      },
                      "revision": 3
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌已失效",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "无权限，或超出账号的数据范围",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "对象不存在",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本已变化，请刷新后重试",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-requirements": [
          "IAA-REQ-001",
          "IAA-REQ-003"
        ],
        "x-screens": [
          "user-manage"
        ],
        "x-fields": [
          "username",
          "roles",
          "data_scope",
          "revision"
        ],
        "x-implementation": "implemented",
        "x-reference-implementation": {
          "runtime": "FastAPI",
          "status": "planned"
        },
        "x-implementation-source": "backend-java/src/main/java/com/iaaops/iam/web/UserAdminController.java"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DataScope": {
        "properties": {
          "agencies": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agencies",
            "description": "可见代理；null 或省略表示不限；[] 表示无授权值"
          },
          "products": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Products",
            "description": "可见产品；null 或省略表示不限；[] 表示无授权值"
          },
          "operators": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operators",
            "description": "可见运营；null 或省略表示不限；[] 表示无授权值"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DataScope",
        "description": "Java 服务端授权范围；维度之间取交集。省略或 null 表示该维度不限，空数组 [] 表示该维度没有任何授权值，查询不会返回事实数据。这是授权范围，不是报表请求中用于清除筛选条件的空筛选数组。"
      },
      "Error": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "about:blank"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "code": {
            "type": "string",
            "enum": [
              "AUTH_REQUIRED",
              "INVALID_CREDENTIALS",
              "PASSWORD_CHANGE_REQUIRED",
              "FORBIDDEN",
              "NOT_FOUND",
              "METHOD_NOT_ALLOWED",
              "CONFLICT",
              "PAYLOAD_TOO_LARGE",
              "VALIDATION_FAILED",
              "RESULT_TOO_LARGE",
              "INTERNAL_ERROR",
              "NOT_IMPLEMENTED",
              "HTTP_ERROR"
            ],
            "title": "Code",
            "x-code-descriptions": {
              "AUTH_REQUIRED": "未登录、令牌缺失或已失效，需要重新登录。",
              "CONFLICT": "对象已被他人修改（revision 不一致）或唯一键冲突，刷新后重试。",
              "FORBIDDEN": "没有该操作的权限，或请求超出账号的数据范围。",
              "HTTP_ERROR": "其他 HTTP 错误，以 status 为准。",
              "INTERNAL_ERROR": "服务端异常，已记录，可稍后重试。",
              "INVALID_CREDENTIALS": "账号或口令错误，或登录暂被限制；不区分账号不存在、口令错误、账号停用与锁定。",
              "METHOD_NOT_ALLOWED": "路径存在，但不支持该 HTTP 方法。",
              "NOT_FOUND": "对象或路径不存在。",
              "NOT_IMPLEMENTED": "契约已确定，接口尚未实现。",
              "PASSWORD_CHANGE_REQUIRED": "正在使用临时口令，修改口令前只能访问当前账号与改密接口。",
              "PAYLOAD_TOO_LARGE": "上传文件超过大小限制。",
              "RESULT_TOO_LARGE": "分组数或导出行数超过上限，请缩小日期范围或减少分组维度。",
              "VALIDATION_FAILED": "请求参数未通过校验，errors 列出具体字段。"
            }
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "status",
          "code"
        ],
        "title": "Error",
        "description": "RFC 9457 Problem Details（application/problem+json），所有 4xx/5xx 统一使用。"
      },
      "FieldError": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "title": "FieldError"
      },
...
      "UserCreate": {
        "properties": {
          "username": {
            "type": "string",
            "pattern": "^[a-z0-9_.-]{3,32}$",
            "title": "Username"
          },
          "display_name": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Display Name"
          },
          "roles": {
            "items": {
              "type": "string",
              "enum": [
                "super_admin",
                "company_admin",
                "operator",
                "agency_admin",
                "customer",
                "readonly"
              ]
            },
            "type": "array",
            "minItems": 1,
            "title": "Roles"
          },
          "data_scope": {
            "$ref": "#/components/schemas/DataScope"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "username",
          "display_name",
          "roles"
        ],
        "title": "UserCreate"
      },
      "UserCreated": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserSummary"
          },
          "one_time_password": {
            "type": "string",
            "title": "One Time Password",
            "description": "仅在本响应返回一次；首次登录必须修改"
          },
          "must_change_password": {
            "type": "boolean",
            "const": true,
            "title": "Must Change Password",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "user",
          "one_time_password"
        ],
        "title": "UserCreated"
      },
      "UserPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UserSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "UserPage"
      },
      "UserRolesUpdate": {
        "properties": {
          "roles": {
            "items": {
              "type": "string",
              "enum": [
                "super_admin",
                "company_admin",
                "operator",
                "agency_admin",
                "customer",
                "readonly"
              ]
            },
            "type": "array",
            "minItems": 1,
            "title": "Roles"
          },
          "data_scope": {
            "$ref": "#/components/schemas/DataScope"
          },
          "revision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "roles",
          "data_scope",
          "revision"
        ],
        "title": "UserRolesUpdate"
      },
      "UserSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "roles": {
            "items": {
              "type": "string",
              "enum": [
                "super_admin",
                "company_admin",
                "operator",
                "agency_admin",
                "customer",
                "readonly"
              ]
            },
            "type": "array",
            "title": "Roles"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "disabled"
            ],
            "title": "Status"
          },
          "data_scope": {
            "$ref": "#/components/schemas/DataScope"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "username",
          "display_name",
          "roles",
          "status",
          "data_scope",
          "revision"
        ],
        "title": "UserSummary"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "本机 Java 业务服务",
      "x-environment": "development"
    },
    {
      "url": "http://127.0.0.1:18081",
      "description": "本机 Java 隔离验收示例，按需启动，非常驻入口",
      "x-environment": "testing"
    },
    {
      "url": "https://iaa-ops.invalid",
      "description": "公开演示规划占位，尚未部署；发布后替换为真实地址",
      "x-environment": "production"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}